Rename GenerationServiceService to GenerationService
